+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha26] - 2019-11-14
+ ### Change
+ - Using Msgpack extension is not default from now.
+ - Checking Msgpack extension is separated to one place.
+ ### Add
+ - Content-ID to header for debugging.
+
 ## [1.0.0-alpha25] - 2019-07-17
  ### Fix
  - Fixing the exception handling during the response parse.